+++ NOTI_Teacher/NOTI_Teacher/Screen/Home/ViewController/DisturbTimeVC.swift
@@ -22,30 +22,10 @@ class DisturbTimeVC: BaseViewController {
             $0.configureStatusMessage(nil)
         }
     
-    private let baseStackView = UIStackView()
-        .then {
-            $0.axis = .vertical
-            $0.layer.cornerRadius = 5
-            $0.layer.borderWidth = 1
-            $0.layer.borderColor = UIColor.lineGray.cgColor
-        }
-    
-    private var startTimeView = TimePickerView()
-        .then {
-            $0.configureTitle("시작 시간")
-            $0.configurePicker(mode: .dateAndTime)
-        }
-    
-    private var endTimeView = TimePickerView()
-        .then {
-            $0.configureTitle("종료 시간")
-            $0.configurePicker(mode: .dateAndTime)
-        }
-    
-    private let separatorView = UIView()
-        .then {
-            $0.backgroundColor = .lineGray
-        }
+    private let timePickerView = TimePickerView(startTitle: "시작 시간",
+                                                endTitle: "종료 시간",
+                                                startPickerMode: .time,
+                                                endPickerMode: .time)
     
     private let bag = DisposeBag()
     
@@ -89,10 +69,7 @@ extension DisturbTimeVC {
     
     private func configureContentView() {
         view.addSubviews([alarmToggleView,
-                          baseStackView])
-        [startTimeView, separatorView, endTimeView].forEach {
-            baseStackView.addArrangedSubview($0)
-        }
+                          timePickerView])
     }
 }
 
@@ -107,27 +84,13 @@ extension DisturbTimeVC {
             $0.height.equalTo(64)
         }
         
-        baseStackView.snp.makeConstraints {
+        timePickerView.snp.makeConstraints {
             $0.top.equalTo(alarmToggleView.snp.bottom).offset(4)
             $0.leading.equalToSuperview().offset(20)
             $0.trailing.equalToSuperview().offset(-20)
         }
         
-        [startTimeView, endTimeView].forEach {
-            $0.snp.makeConstraints {
-                $0.height.equalTo(62)
-            }
-        }
-        
-        separatorView.snp.makeConstraints {
-            $0.height.equalTo(1)
-        }
-        
-        baseStackView.isLayoutMarginsRelativeArrangement = true
-        baseStackView.layoutMargins = UIEdgeInsets(top: .zero,
-                                                   left: 20,
-                                                   bottom: .zero,
-                                                   right: 20)
+        timePickerView.sizeToFit()
     }
 }
 
@@ -139,8 +102,7 @@ extension DisturbTimeVC {
             .asDriver()
             .drive(onNext: {[weak self] isOn in
                 guard let self = self else { return }
-                // TODO: - 비활성화, 활성화 디자인 적용
-                print(isOn)
+                self.timePickerView.isUserInteractionEnabled = isOn
             })
             .disposed(by: bag)
     }
